API Reference

CopernicusData.get_AWS_config — Function

function get_AWS_config(profile::String="default")

Generate and AWS config given a profile name. To get the credentials, the function follows the priority as: 1- get aws config and credentials file as INI files in default path: ~/.aws 2- use environement varialbes (AWSSECRETACCESSKEY, AWSACCESSKEYID, ...)

CopernicusData.EOProducts.eoproduct_dataset — Method

eoproduct_dataset(path::String)::Dict{String, Dataset}

Open a Copernicus product returns a Dict{String,Dataset} containing all the Variables stored in the product

Examples

julia> d = open_eoproduct("S3SLSLST_20191227T124111_0179_A109_T921.zarr")

CopernicusData.EOProducts.open_eoproduct — Method

open_eoproduct(path::String)

Open a Copernicus Zarr product returns a Dict{String,ZArray} containing all the Variables stored in the product

Examples

julia> d = open_eoproduct("S3SLSLST_20191227T124111_0179_A109_T921.zarr")

Base.iterate — Method

Base.iterate(tree::YAXTree, state::YAXTreeIteratorState)

Base.iterate — Method

Base.iterate(tree::YAXTree)

Creates an iterator that traverses the YAXTree in a depth-first, pre-order fashion.

This iterator yields each node in the tree, ensuring that no node is visited twice, even if children share the same name. The path of each node is used as a unique identifier.

Examples

# Assuming a YAXTree 'my_tree' is defined:
for node in my_tree
    println(node.name, " at path ", node.path)
end

CopernicusData.YAXTrees.open_datatree — Function

open_datatree(path::String, driver::Union{Nothing,Symbol}=nothing; name::String="root")::YAXTree

Open a Copernicus product returns YAXTree

Examples

julia> dt = open_datatree("S03SLSLST_20191227T124111_0179_A109_T921.zarr")

CopernicusData.YAXTrees.@map_over_subtrees — Macro

macro map_over_subtrees(func, tree)

Map a function func over all the nodes of a YAXTree structure which contains a YAXArray or YAXArrays.Datasets.Dataset

Examples

# Assuming a YAXTree 'my_tree' is defined:
f(tree::YAXTree) = @show tree.name
YAXTrees.@map_over_subtrees f my_tree